# docs/04-api/openapi/frameflow-v1.yaml
openapi: 3.0.3
info:
  title: FrameFlow Select API
  version: 1.0.0
  description: FrameFlow Select 批量质检与 Top-K 优选平台。身份端点沿用 M01（/api/v1/auth、/api/v1/teams）；产品端点为手写 OpenAPI 权威契约，运行时以 springdoc 输出并经验证对齐。
...
paths:
...
  /api/v1/selection-sets/{selectionSetId}/lock:
    post:
      tags:
      - product
      summary: 锁定选择集（不可变）
      operationId: lockSelectionSet
      parameters:
      - name: selectionSetId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: X-Team-Id
        in: header
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SelectionSet'
      security:
      - bearerAuth: []
...
  /api/v1/batches/{batchId}/similarity-clusters:
    get:
      tags:
      - product
      summary: 批次聚类列表
      operationId: listSimilarityClusters
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimilarityCluster'
      security:
      - bearerAuth: []
    post:
      tags:
      - product
      summary: 计算候选相似度聚类
      operationId: createSimilarityClusters
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterRequest'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimilarityCluster'
      security:
      - bearerAuth: []
  /api/v1/batches/{batchId}/selection-sets:
    get:
      tags:
      - product
      summary: 批次选择集列表
      operationId: listSelectionSets
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SelectionSet'
      security:
      - bearerAuth: []
    post:
      tags:
      - product
      summary: 创建选择集（Top-K）
      operationId: createSelectionSet
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSelectionRequest'
        required: true
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SelectionSet'
      security:
      - bearerAuth: []
  /api/v1/batches/{batchId}/ranking-snapshots:
    get:
      tags:
      - product
      summary: 批次排名快照列表
      operationId: listRankingSnapshots
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: {}
      security:
      - bearerAuth: []
    post:
      tags:
      - product
      summary: 生成排名快照
      operationId: createRankingSnapshot
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RankingSnapshot'
      security:
      - bearerAuth: []
...
  /api/v1/selection-sets/{selectionSetId}/export:
    get:
      tags:
      - product
      summary: 导出选择集（CSV/JSON）
      operationId: exportSelectionSet
      parameters:
      - name: selectionSetId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: X-Team-Id
        in: header
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
      security:
      - bearerAuth: []
  /api/v1/ranking-snapshots/{snapshotId}:
    get:
      tags:
      - product
      summary: 排名快照（含分值构成）
      operationId: getRankingSnapshot
      parameters:
      - name: snapshotId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RankingEntry'
      security:
      - bearerAuth: []
...
  /api/v1/batches/{batchId}:
    get:
      tags:
      - product
      summary: 获取批次
      operationId: getBatch
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Batch'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    SelectionSet:
      type: object
      properties:
        id:
          type: integer
          format: int64
        batchId:
          type: integer
          format: int64
        lockedBy:
          type: integer
          format: int64
        createdBy:
          type: integer
          format: int64
        name:
          type: string
        status:
          type: string
        theme:
          type: string
        topK:
          type: integer
          format: int32
        lockedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ClusterRequest:
      type: object
      properties:
        threshold:
          type: number
          format: double
    SimilarityCluster:
      type: object
      properties:
        id:
          type: integer
          format: int64
        batchId:
          type: integer
          format: int64
        representativeCandidateId:
          type: integer
          format: int64
        clusterKey:
          type: string
        algorithmVersion:
          type: string
        threshold:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
    CreateSelectionRequest:
      type: object
      properties:
        name:
          type: string
        topK:
          type: integer
          format: int32
    RankingSnapshot:
      type: object
      properties:
        id:
          type: integer
          format: int64
        batchId:
          type: integer
          format: int64
        createdBy:
          type: integer
          format: int64
        snapshotVersion:
          type: integer
          format: int32
        algorithmVersion:
          type: string
        weights:
          type: string
        featureSchemaVersion:
          type: string
        lockedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    RankingEntry:
      type: object
      properties:
        rankingSnapshotId:
          type: integer
          format: int64
        candidateId:
          type: integer
          format: int64
        rank:
          type: integer
          format: int32
        baseScore:
          type: number
          format: double
        finalScore:
          type: number
          format: double
        breakdown:
          type: string
...
    Batch:
      type: object
      properties:
        id:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        name:
          type: string
        status:
          type: string
        promptText:
          type: string
        profileVersionId:
          type: integer
          format: int64
        briefId:
          type: integer
          format: int64
        capacityLimit:
          type: integer
          format: int32
        failureCount:
          type: integer
          format: int32
        createdBy:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  headers:
    RequestId:
      description: 服务器为本次 HTTP 请求生成的请求 ID
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT